Very basic collision detection system
Projectile from ship's cannon will now collide and delete the projectile
and the collided asteroid.

The empty alpha background padding around the asteroid needs to
be accounted for next to ensure the project visually touches the
asteroid on collision before they're deleted.

Need to also think of a better architecture to get access to the bounds
of each project and managing destroying them efficiently.